requirements.txt
# Core Backend 
Django>=5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt  
python-dotenv

# Data Science & PDF 
pandas
numpy
xhtml2pdf


# Desktop Application 
PyQt5
requests
matplotlib

# Utilities
certifi
urllib3